[BY-239] Move: create folders to organize widget
