+++ lib/screen/authenticate/info/setup_user.dart
@@ -2,12 +2,12 @@ import 'package:byourside/screen/authenticate/controller/nickname_controller.dar
 import 'package:byourside/model/authenticate/save_user_data.dart';
 import 'package:byourside/utils/validation_utils.dart';
 import 'package:byourside/widget/app_bar.dart';
-import 'package:byourside/widget/authenticate/age_section.dart';
-import 'package:byourside/widget/authenticate/disability_type_button.dart';
-import 'package:byourside/widget/authenticate/location_section.dart';
-import 'package:byourside/widget/authenticate/nickname_section.dart';
-import 'package:byourside/widget/authenticate/purpose_select.dart';
-import 'package:byourside/widget/authenticate/user_type_button.dart';
+import 'package:byourside/widget/authenticate/setup/age_section.dart';
+import 'package:byourside/widget/authenticate/setup/disability_type_button.dart';
+import 'package:byourside/widget/authenticate/setup/location_section.dart';
+import 'package:byourside/widget/authenticate/setup/nickname_section.dart';
+import 'package:byourside/widget/authenticate/setup/purpose_select.dart';
+import 'package:byourside/widget/authenticate/setup/user_type_button.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
