+++ lib/widget/authenticate/setup/nickname_section.dart
